Finish merging
