+++ src/App.js
@@ -1,28 +1,24 @@
 import { Routes, Route } from "react-router-dom"
 import Homepage from "./Pages/Homepage";
 import Header from "./Components/Header";
-import Navbar from "./Components/Navbar";
 import Comparepage from "./Pages/Comparepage";
-import SignIn from "./Pages/SignIn/SignIn";
-import SignUp from "./Pages/SignUp/SignUp";
+import SignIn from "./Pages/SignIn";
+import SignUp from "./Pages/SignUp";
 import Footer from "./Components/Footer";
-import "./Components/websiteTheme.css";
+import "./websiteTheme.css";
 
 const App = () => {
   return (
-    <div className="wrapper">
-      <div className="static">
-        <Header />
-      </div>
-      <Navbar />
+    <div>
+      <Header />
       <div className="content">
-      <Routes>
-        <Route path="/" element={<Homepage />} />
-        <Route path="/signin" element={<SignIn />} />
-        <Route path="/signup" element={<SignUp />} />
-        <Route path="/compare" element={<Comparepage />} />
-      </Routes>
-      <Footer />
+        <Routes>
+          <Route path="/" element={<Homepage />} />
+          <Route path="/signin" element={<SignIn />} />
+          <Route path="/signup" element={<SignUp />} />
+          <Route path="/compare" element={<Comparepage />} />
+        </Routes>
+        <Footer />
       </div>
     </div>
   );
