+++ src/Components/Footer.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Link } from "react-router-dom";
 import "./footer.css";
-import "./websiteTheme.css";
+import "../websiteTheme.css";
 
 const Footer = () => {
   return (
