+++ src/Components/header.css
@@ -1,4 +1,4 @@
-.header {
+.headerContent {
   box-sizing: border-box;
   width: 100%;
   height: 100px;
@@ -6,85 +6,45 @@
   flex: 4em 1em 1em;
   align-items: flex-end;
   color: black;
+  background-color: white;
+}
+
+.header {
+  position: fixed;
+  top: 0;
+  width: 100%;
+  z-index: 2;
 }
 
 .pageName {
-  font-family: 'hobo-std';
   margin-right: auto;
-  align-items: center;
-  background-color: #E6E6E6;
-  color: #000000;
-  padding: 1em;
-  padding-left: 2em;
 }
 
-.headerText {
-  align-items: center;
-  font-family: 'Open Sans';
-  font-style: normal;
-  font-weight: 600;
-  font-size: 16px;
-  line-height: 33px;
-  color: #000000;
-  margin: auto;
-}
-
-.aboveHeader {
-  box-sizing: border-box;
-  width: 100%;
-  background-color: #FFFFFF;
+.logoWithText {
   display: flex;
   color: #0FA958;
+  align-items: center;
   align-items: flex-end;
   text-decoration: none;
   margin-right: 1em;
-  margin-bottom: 0.2em;
+  padding-bottom: 0.2em;
   font-size: x-large;
-  flex: 4em 1em 1em;
+}
+
+.navbar {
+  background-color: lightgray;
+  height: 70px;
+  display: flex;
+  justify-content: space-between;
   align-items: center;
-  padding: 0.5em;
   padding-left: 1em;
+  padding-right: 1em;
 }
 
-.aboveHeaderLeft {
-  float: left;
-  margin-right: auto;
-}
-
-.aboveHeaderRight {
-  float: right;
-  margin-left: auto;
-}
-
-.aboveHeaderItems {
-  float: left;
-  align-items: center;
-  white-space: nowrap;
-  padding-right: 2em;
-}
-
-.avoidCollapse {
-  clear: left;
-}
-
-.aboveHeaderSingleText {
-  display: inline-block;
-  font-family: 'Open Sans';
-  font-size: 18px;
-  color: #0FA958;
+.navbaritem {
   text-decoration: none;
-}
-
-.otherHeaderLogos {
-  height: auto;
-  width: auto;
-  max-width: 50px;
-}
-
-.mainLogo {
-  height: auto;
-  width: auto;
-  max-width: 200px;
+  color: black;
+  font-size: x-large;
 }
 
 .inputToCompare {
