--- src/Components/navbar.css
@@ -1,13 +1,0 @@
-.navbar {
-  background-color: lightgray;
-  height: 70px;
-  display: flex;
-  justify-content: space-evenly;
-  align-items: center;
-}
-
-.navbaritem {
-  text-decoration: none;
-  color: black;
-  font-size: x-large;
-}
