+++ src/Pages/Comparepage/index.js
@@ -1,3 +1,3 @@
 import Comparepage from "./Comparepage";
 
-export default Comparepage
+export default Comparepage;
