+++ src/Pages/Homepage/index.js
@@ -1,0 +1,3 @@
+import Homepage from "./Homepage";
+
+export default Homepage;
