+++ src/Pages/SignIn/index.js
@@ -1,0 +1,3 @@
+import SignIn from "./SignIn";
+
+export default SignIn;
