+++ src/websiteTheme.css
@@ -1,0 +1,37 @@
+a:visited {
+    text-decoration: none;
+    color: #000000;
+}
+
+a:hover {
+    text-decoration: none;
+    color: #000000;
+}
+
+a:focus {
+    text-decoration: none;
+    color: #000000;
+}
+
+a:active {
+    text-decoration: none;
+    color: #000000;
+}
+
+input {
+    text-align: center;
+}
+
+::placeholder {
+    /* Chrome, Firefox, Opera, Safari 10.1+ */
+    color: #0FA958;
+    font-weight: 550;
+    font-style: italic;
+}
+
+.content {
+    float: left;
+    position: relative;
+    top: 170px;
+    width: 100%;
+}
